requirements.txt
# Core RAG components
chromadb==0.4.18
sentence-transformers==2.2.2
huggingface-hub==0.16.4
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utilities
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3